# Parallelize all the tests using ForAllNets (cut test time by half)