@elichai 2020-12-08
As a side effect this also allows you to run a test only for a specific network e.g. `go test -run TestDifficulty/kaspa-mainnet`