@elichai 2020-12-08
The only downside is that it seems that the tests will now take ~2GB of RAM instead of ~980MB